move helper validation component into common AppEngineDeploymentConfigurationPanel.java, remove identical update code.
